sync: hosted cloud (#1540)
Submit a pull request for this project.

<!-- If you have an Issue that related to this Pull Request, you can
copy this Issue's description -->

# Why? 
<!-- 
> Related to which issue?
> Why we need this pull request?
> What is the user story for this pull request? 
-->


# What?
<!-- 
> Can you describe this feature in detail?
> Who can benefit from it? 
-->


# How?
<!-- 
> Do you have a simple description of how this pull request is
implemented?
-->

---------

+++ backend-server/application/src/main/java/com/apitable/workspace/service/impl/ControlMemberServiceImpl.java
@@ -93,16 +93,16 @@ public class ControlMemberServiceImpl implements IControlMemberService {
         Map<Long, ControlMemberDTO> memberControlRoleMap =
             this.getMemberControlRoleMap(spaceId, controlId);
         // calculate position
-        int sub = (int) (page.getCurrent() - 1) * (int) page.getSize();
+        long sub = (page.getCurrent() - 1) * page.getSize();
         if (sub > memberControlRoleMap.size()) {
-            return PageHelper.build((int) page.getCurrent(), (int) page.getSize(),
+            return PageHelper.build(page.getCurrent(), page.getSize(),
                 memberControlRoleMap.size(), new ArrayList<>());
         }
 
-        int end = (sub + page.getSize()) > memberControlRoleMap.size()
-            ? memberControlRoleMap.size() : sub + (int) page.getSize();
+        long end = (sub + page.getSize()) > memberControlRoleMap.size()
+            ? memberControlRoleMap.size() : sub + page.getSize();
         List<Long> memberIds =
-            new ArrayList<>(memberControlRoleMap.keySet()).subList(sub, end);
+            new ArrayList<>(memberControlRoleMap.keySet()).subList((int) sub, (int) end);
         List<T> records = new ArrayList<>(memberIds.size());
         List<NodeRoleMemberVo> results = iMemberService.getNodeRoleMemberWithSort(memberIds);
         // Give permission value
@@ -113,7 +113,7 @@ public class ControlMemberServiceImpl implements IControlMemberService {
             result.setIsControlOwner(controlMemberDTO.getIsControlOwner());
             records.add(BeanUtil.toBean(result, clz));
         });
-        return PageHelper.build((int) page.getCurrent(), (int) page.getSize(),
+        return PageHelper.build(page.getCurrent(), page.getSize(),
             memberControlRoleMap.size(), records);
     }
